split wallet and verifier metadata
