+++ src/oid4vp/wallet.rs
@@ -1,7 +1,9 @@
-//! #! Querying credentials
+//! # Wallet 
 
 mod response;
 mod vp_token;
+mod metadata;
 
 pub use response::*;
 pub use vp_token::*;
+pub use metadata::*;
